Move output function to the right file
