+++ include/memory_map.h
@@ -7,9 +7,8 @@
 
 /*Output functions*/
 void create_object_file();
-
 void create_entry_file();
-
 void create_external_file();
+void create_output_files(int* line_counter);
 
 #endif
