+++ src/second_pass.c
@@ -3,8 +3,7 @@
 #include "../include/first_pass.h"
 #include "../include/second_pass.h"
 #include "../include/logger.h"
-
-
+#include "memory_map.h"
 
 
 /**
